Extract ConsoleOutput and ConsoleToggle components to reduce App.vue complexity
- Create ConsoleOutput.vue for overlay display functionality
- Create ConsoleToggle.vue for toggle button with clean interface
- Remove confusing toggleOnly/overlayOnly props abstraction
- Clear separation: toggle shows state, output manages overlay
- Extract all console-related logic and styling from App.vue
- Reduce App.vue by ~100 lines with better separation of concerns
