+++ src/App.vue
@@ -6,6 +6,8 @@ import ExampleSelector from './components/ExampleSelector.vue'
 import OutputDisplay from './components/OutputDisplay.vue'
 import WebRStatus from './components/WebRStatus.vue'
 import LibrarySelector from './components/LibrarySelector.vue'
+import ConsoleOutput from './components/ConsoleOutput.vue'
+import ConsoleToggle from './components/ConsoleToggle.vue'
 import { useWebR } from './composables/useWebR'
 import { examples } from './data/examples'
 import { icons } from './data/icons'
@@ -50,47 +52,8 @@ const {
   toggleLibrary,
 } = useWebR()
 
-// Filter text messages for console display
-const textMessages = computed(() => {
-  return messages.filter(message => message.type !== 'plot')
-})
-
-// Console state
-const showConsole = ref(false)
-const hasErrors = computed(() => {
-  return textMessages.value.some(msg => msg.type === 'error' || msg.type === 'stderr')
-})
-
-const hasWarnings = computed(() => {
-  return textMessages.value.some(msg => msg.type === 'warning')
-})
-
-// Group consecutive messages of the same type for cleaner display
-const groupedMessages = computed(() => {
-  const groups: Array<{type: string, content: string, showLabel: boolean}> = []
-  
-  for (const message of textMessages.value) {
-    const lastGroup = groups[groups.length - 1]
-    
-    // Clean up excessive line breaks from the message content
-    const cleanContent = message.content.replace(/\n\s*\n\s*\n/g, '\n').trim()
-    
-    if (lastGroup && lastGroup.type === message.type) {
-      // Same type as previous - append content with single line break
-      lastGroup.content += '\n' + cleanContent
-    } else {
-      // New type - create new group
-      const showLabel = message.type !== 'stdout' // Don't show label for stdout
-      groups.push({
-        type: message.type,
-        content: cleanContent,
-        showLabel
-      })
-    }
-  }
-  
-  return groups
-})
+// Console component reference
+const consoleRef = ref<InstanceType<typeof ConsoleOutput>>()
 
 const runCode = async (): Promise<void> => {
   if (code.value.trim()) {
@@ -99,10 +62,7 @@ const runCode = async (): Promise<void> => {
     lastExecutedCode.value = code.value
     
     // Auto-open console if no plots were generated
-    const hasPlots = messages.some(msg => msg.type === 'plot')
-    if (!hasPlots && messages.length > 0) {
-      showConsole.value = true
-    }
+    consoleRef.value?.autoOpenConsole()
   }
 }
 
@@ -162,10 +122,7 @@ const handleExampleSelect = async (example: RExample): Promise<void> => {
           lastExecutedCode.value = example.code
           
           // Auto-open console if no plots were generated
-          const hasPlots = messages.some(msg => msg.type === 'plot')
-          if (!hasPlots && messages.length > 0) {
-            showConsole.value = true
-          }
+          consoleRef.value?.autoOpenConsole()
         }
       }
     } catch (error) {
@@ -184,10 +141,7 @@ const handleExampleSelect = async (example: RExample): Promise<void> => {
       lastExecutedCode.value = example.code
       
       // Auto-open console if no plots were generated
-      const hasPlots = messages.some(msg => msg.type === 'plot')
-      if (!hasPlots && messages.length > 0) {
-        showConsole.value = true
-      }
+      consoleRef.value?.autoOpenConsole()
     }
   }
 }
@@ -306,44 +260,10 @@ onMounted(async () => {
             :messages="messages"
             :is-loading="isLoading"
           />
-          <!-- Console overlay when expanded -->
-          <div
-            v-if="showConsole"
-            class="console-overlay"
-          >
-            <div class="console-content">
-              <div class="console-header">
-                <span class="console-title">Console Output</span>
-                <button
-                  class="console-close"
-                  @click="showConsole = false"
-                >
-                  ×
-                </button>
-              </div>
-              <div class="console-messages">
-                <div 
-                  v-for="(group, index) in groupedMessages" 
-                  :key="'group-' + index" 
-                  class="console-message" 
-                  :class="[group.type, { 'no-label': !group.showLabel }]"
-                >
-                  <span
-                    v-if="group.showLabel"
-                    class="message-label"
-                  >{{ group.type.toUpperCase() }}:</span>
-                  <pre 
-                    v-if="group.type === 'stdout' || group.type === 'stderr'" 
-                    class="message-text"
-                  >{{ group.content }}</pre>
-                  <span
-                    v-else
-                    class="message-text"
-                  >{{ group.content }}</span>
-                </div>
-              </div>
-            </div>
-          </div>
+          <ConsoleOutput 
+            ref="consoleRef"
+            :messages="messages"
+          />
         </div>
       </div>
 
@@ -366,17 +286,11 @@ onMounted(async () => {
           </div>
         </div>
         <div class="bottom-bar-right">
-          <button 
-            class="console-toggle"
-            :class="{ 'has-errors': hasErrors, 'has-warnings': hasWarnings && !hasErrors }"
-            @click="showConsole = !showConsole"
-          >
-            Console ({{ textMessages.length }})
-            <span
-              class="toggle-arrow"
-              :class="{ 'open': showConsole }"
-            >▼</span>
-          </button>
+          <ConsoleToggle 
+            :messages="messages"
+            :is-open="consoleRef?.showConsole || false"
+            @toggle="consoleRef?.toggleConsole"
+          />
         </div>
       </div>
     </main>
@@ -558,119 +472,6 @@ onMounted(async () => {
   position: relative;
 }
 
-.console-overlay {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background: white;
-  border: 1px solid #e5e7eb;
-  border-radius: 6px 6px 0 0;
-  box-shadow: 0 -2px 8px rgba(0, 0, 0, 0.1);
-  z-index: 10;
-  max-height: 60%;
-}
-
-.console-content {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.console-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem 1rem;
-  border-bottom: 1px solid #e5e7eb;
-  background: #f9fafb;
-}
-
-.console-title {
-  font-weight: 600;
-  font-size: 0.875rem;
-  color: #374151;
-}
-
-.console-close {
-  background: none;
-  border: none;
-  font-size: 1.25rem;
-  cursor: pointer;
-  color: #6b7280;
-  padding: 0;
-  width: 1.5rem;
-  height: 1.5rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 4px;
-}
-
-.console-close:hover {
-  background: #e5e7eb;
-  color: #374151;
-}
-
-.console-messages {
-  flex: 1;
-  overflow-y: auto;
-  padding: 0.75rem;
-  min-height: 120px;
-  max-height: 200px;
-}
-
-.console-message {
-  display: flex;
-  gap: 0.5rem;
-  margin-bottom: 0.5rem;
-  font-size: 0.75rem;
-  line-height: 1.4;
-}
-
-.console-message.no-label {
-  gap: 0;
-}
-
-.console-message:last-child {
-  margin-bottom: 0;
-}
-
-.message-label {
-  font-weight: 600;
-  text-transform: uppercase;
-  color: #6b7280;
-  min-width: 3rem;
-  flex-shrink: 0;
-}
-
-.console-message.success .message-label {
-  color: #059669;
-}
-
-.console-message.error .message-label {
-  color: #dc2626;
-}
-
-.console-message.stderr .message-label {
-  color: #d97706;
-}
-
-.console-message.info .message-label {
-  color: #2563eb;
-}
-
-.console-message.warning .message-label {
-  color: #d97706;
-}
-
-.message-text {
-  flex: 1;
-  margin: 0;
-  font-family: 'Courier New', monospace;
-  color: #374151;
-  white-space: pre-wrap;
-}
 
 /* Bottom bar styling */
 .bottom-bar {
@@ -708,45 +509,6 @@ onMounted(async () => {
   align-items: center;
 }
 
-.console-toggle {
-  background: #f3f4f6;
-  border: 1px solid #d1d5db;
-  border-radius: 6px;
-  padding: 0.5rem 0.75rem;
-  font-size: 0.875rem;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  color: #6b7280;
-}
-
-.console-toggle:hover {
-  background: #e5e7eb;
-  border-color: #9ca3af;
-}
-
-.console-toggle.has-errors {
-  background: #fef2f2;
-  border-color: #fecaca;
-  color: #dc2626;
-}
-
-.console-toggle.has-warnings {
-  background: #fefbf2;
-  border-color: #fed7aa;
-  color: #d97706;
-}
-
-.toggle-arrow {
-  font-size: 0.75rem;
-  transition: transform 0.3s ease;
-}
-
-.toggle-arrow.open {
-  transform: rotate(180deg);
-}
 
 
 @media (max-width: 768px) {
